telemetry: Amazon Q indicate if it can auto connect
In the telemetry event auth_userState we need to indicate
our authStatus field as 'connected' if the Amazon Q extension does
not have a connection, but we detect one exists already in the
AWS Toolkit extension.

This commit adds extra logic which checks if an autoconnectable
connection can be found in AWS Toolkit and then indicates that we
are connected. This assumes that the separate Login auto-connect logic
will eventually do the auto connection.
