+++ packages/core/src/codewhisperer/util/authUtil.ts
@@ -429,6 +429,21 @@ export class AuthUtil {
 
         return state
     }
+
+    /**
+     * From the given connections, returns a connection that works with Amazon Q.
+     */
+    findUsableQConnection(connections: AwsConnection[]): AwsConnection | undefined {
+        const hasQScopes = (c: AwsConnection) => amazonQScopes.every(s => c.scopes?.includes(s))
+        const score = (c: AwsConnection) => Number(hasQScopes(c)) * 10 + Number(c.state === 'valid')
+        connections.sort(function (a, b) {
+            return score(b) - score(a)
+        })
+        if (hasQScopes(connections[0])) {
+            return connections[0]
+        }
+        return undefined
+    }
 }
 
 /**
