+++ packages/core/src/shared/index.ts
@@ -15,3 +15,4 @@ export { RegionProvider } from './regions/regionProvider'
 export { Commands } from './vscode/commands2'
 export { getMachineId } from './vscode/env'
 export { getLogger } from './logger/logger'
+export { activateExtension } from './utilities/vsCodeUtils'
